feat: add dynamic rendering and email verification improvements
- Force dynamic rendering for multiple dashboard pages
- Extract email verification logic into separate component
- Add TypeScript ignore comments for type safety
- Simplify search conversations dialog props
- Add video chat stream type safety
