+++ src/app/(dashboards)/(general)/feedbacks/page.tsx
@@ -1,4 +1,5 @@
 'use client';
+export const dynamic = 'force-dynamic';
 
 import React, { useState } from 'react';
 import { useFeedbackList } from '@/hooks/useFeedback';
