+++ src/app/(dashboards)/(general)/templates/page.tsx
@@ -1,4 +1,5 @@
 'use client';
+export const dynamic = 'force-dynamic';
 
 import { useState, useEffect } from 'react';
 import { Scale, Search, Plus } from 'lucide-react';
