+++ src/app/(dashboards)/lawyers/page.tsx
@@ -1,3 +1,4 @@
+export const dynamic = 'force-dynamic';
 import LayersPageContent from '@/components/dashboard/lawyers/LayersPageContent';
 
 export default async function LawyersPage() {
