+++ src/app/(dashboards)/users/page.tsx
@@ -1,4 +1,5 @@
 import UserPageContent from '@/components/dashboard/users/UserPageContent';
+export const dynamic = 'force-dynamic';
 
 export default async function UsersPage() {
   return <UserPageContent />;
