+++ src/app/profile/page.tsx
@@ -1,4 +1,6 @@
 'use client';
+
+export const dynamic = 'force-dynamic';
 import React from 'react';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/Tabs';
 import UpdateProfileForm from '@/components/profile/UpdateProfileForm';
